backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==11.3.0
python-dotenv==1.0.0
google-generativeai==0.3.2
django-filter==23.5 